+++ Aqua/inc/forces.hpp
@@ -6,10 +6,10 @@
 #include "vec.hpp"
 
 namespace {
-inline float square(float x) {
+inline double square(double x) {
     return x * x;
 }
-inline float square_signed(float x) {
+inline double square_signed(double x) {
     return std::abs(x) * x;
 }
 }  // namespace
@@ -18,13 +18,13 @@ namespace aq {
 
 class SeparationForce : public Force {
    protected:
-    const float safeDist;
+    const double safeDist;
 
    public:
-    explicit SeparationForce(float weight, float safe_distance) : Force(weight), safeDist(safe_distance){};
+    explicit SeparationForce(double weight, double safe_distance) : Force(weight), safeDist(safe_distance){};
     virtual void accum(const Fish &near) {
         vec diff = near.getLocation() - me->getLocation();
-        float dist = diff.len();
+        double dist = diff.len();
         if (dist < safeDist) sum -= diff * square(safeDist - dist);
     }
     virtual void finalize() {}
@@ -41,15 +41,15 @@ class AlignmentForce : public Force {
     size_t n_of_close{0};
 
    public:
-    explicit AlignmentForce(float weight) : Force(weight){};
+    explicit AlignmentForce(double weight) : Force(weight){};
     virtual void accum(const Fish &near) {
         sum += near.getVelocity();
         ++n_of_close;
     }
     virtual void finalize() {
         if (n_of_close == 0) return;
-        sum.x /= n_of_close;
-        sum.y /= n_of_close;
+        sum.x /= static_cast<double>(n_of_close);
+        sum.y /= static_cast<double>(n_of_close);
         sum -= me->getVelocity();
         n_of_close = 0;
     }
@@ -66,7 +66,7 @@ class CohesionForce : public Force {
     size_t n_of_close{0};
 
    public:
-    explicit CohesionForce(float weight) : Force(weight){};
+    explicit CohesionForce(double weight) : Force(weight){};
     virtual void accum(const Fish &near) {
         vec diff = near.getLocation() - me->getLocation();
         sum += diff;
@@ -74,8 +74,8 @@ class CohesionForce : public Force {
     }
     virtual void finalize() {
         if (n_of_close == 0) return;
-        sum.x /= n_of_close;
-        sum.y /= n_of_close;
+        sum.x /= static_cast<double>(n_of_close);
+        sum.y /= static_cast<double>(n_of_close);
         n_of_close = 0;
     }
     virtual ~CohesionForce() {}
@@ -91,7 +91,7 @@ class SpeciesCohesionForce : public Force {
     size_t n_of_close{0};
 
    public:
-    explicit SpeciesCohesionForce(float weight) : Force(weight){};
+    explicit SpeciesCohesionForce(double weight) : Force(weight){};
     virtual void accum(const Fish &near) {
         if (!me->sameSpeciesAs(near)) return;
         vec diff = near.getLocation() - me->getLocation();
@@ -100,8 +100,8 @@ class SpeciesCohesionForce : public Force {
     }
     virtual void finalize() {
         if (n_of_close == 0) return;
-        sum.x /= n_of_close;
-        sum.y /= n_of_close;
+        sum.x /= static_cast<double>(n_of_close);
+        sum.y /= static_cast<double>(n_of_close);
         n_of_close = 0;
     }
     virtual ~SpeciesCohesionForce() {}
@@ -114,7 +114,7 @@ class SpeciesCohesionForce : public Force {
 
 class WaterResistanceForce : public Force {
    public:
-    explicit WaterResistanceForce(float weight) : Force(weight){};
+    explicit WaterResistanceForce(double weight) : Force(weight){};
     virtual void accum(const Fish &near) {
         (void)near;
     }
@@ -132,7 +132,7 @@ class WaterResistanceForce : public Force {
 
 class MinSpeedForce : public Force {
    public:
-    explicit MinSpeedForce(float weight) : Force(weight){};
+    explicit MinSpeedForce(double weight) : Force(weight){};
     virtual void accum(const Fish &near) {
         (void)near;
     }
@@ -152,11 +152,11 @@ class MinSpeedForce : public Force {
 
 class MouseForce : public Force {
    protected:
-    const float fearDist;
+    const double fearDist;
     const std::atomic<sf::Vector2f> &mousePosition;
 
    public:
-    explicit MouseForce(float weight, float fear_distance, const std::atomic<sf::Vector2f> &mouse_position) : Force(weight), fearDist(fear_distance), mousePosition(mouse_position){};
+    explicit MouseForce(double weight, double fear_distance, const std::atomic<sf::Vector2f> &mouse_position) : Force(weight), fearDist(fear_distance), mousePosition(mouse_position){};
     virtual void accum(const Fish &near) {
         (void)near;
     }
@@ -164,7 +164,7 @@ class MouseForce : public Force {
         vec mouse = mousePosition.load();
         if (mouse == vec(0, 0)) return;
         vec diff = mouse - me->getLocation();
-        float dist = diff.len();
+        double dist = diff.len();
         if (dist < fearDist) sum -= diff * (fearDist - dist);
     }
     virtual Force *clone() {
@@ -180,17 +180,17 @@ class IslandForce : public Force {
 
     inline constexpr static const size_t nOfSamplePoints = 36;
     inline constexpr static const struct {
-        float x, y;
+        double x, y;
     } samplePoints[nOfSamplePoints] =
         {{1.000, 0.000}, {0.940, 0.342}, {0.766, 0.643}, {0.500, 0.866}, {0.174, 0.985}, {-0.174, 0.985}, {-0.500, 0.866}, {-0.766, 0.643}, {-0.940, 0.342}, {-1.000, 0.000}, {-0.940, -0.342}, {-0.766, -0.643}, {-0.500, -0.866}, {-0.174, -0.985}, {0.174, -0.985}, {0.500, -0.866}, {0.766, -0.643}, {0.940, -0.342}, {0.667, 0.000}, {0.577, 0.333}, {0.333, 0.577}, {0.000, 0.667}, {-0.333, 0.577}, {-0.577, 0.333}, {-0.667, 0.000}, {-0.577, -0.333}, {-0.333, -0.577}, {-0.000, -0.667}, {0.333, -0.577}, {0.577, -0.333}, {0.333, 0.000}, {0.167, 0.289}, {-0.167, 0.289}, {-0.333, 0.000}, {-0.167, -0.289}, {0.167, -0.289}};
 
    public:
-    explicit IslandForce(float weight, const Island::Map &map) : Force(weight), map(map){};
+    explicit IslandForce(double weight, const Island::Map &map) : Force(weight), map(map){};
     virtual void accum(const Fish &near) {
         (void)near;
     }
     virtual void finalize() {
-        float vis = me->getVision();
+        double vis = me->getVision();
         vec loc = me->getLocation();
         size_t land_cnt = 0;
         for (size_t i = 0; i < nOfSamplePoints; i++) {
